Configure "homepage" property in package.json
Summary:
This indicates to `react-scripts` that the application will not be
hosted at the server root, but instead on a GitHub pages subdirectory.

Test Plan:
Add `I am at: {JSON.stringify(process.env.PUBLIC_URL)}` to the `App`
component. Note that in `yarn start` this renders as `""`, while in
`yarn build` it renders as `"/spelling-bee"`.