Add a component for loading word lists
Summary:
This component lists the available word lists, fetches word lists upon
request, and maintains the currently selected word list.

Test Plan:
Change the callback in `App.js` to additionally `console.log` the word
data. Then, load the application, and note that the first word list is
loaded by default, the second word list is loaded on demand, and that
word lists are never re-fetched. Use dev tools to block the URLs for the
first and second word lists, independently and together, and note that
the failures are correctly displayed in the UI. Verify that when the URL
is unblocked, the word list can be re-requested and successfully loaded.
Note that if the construction of the `PuzzleMaster` is mocked to take a
long time (by only `resolve`ing in a `setTimeout`) then the “processing”
stage appears correctly.